+++ integration/spring-data/base/src/main/java/com/blazebit/persistence/spring/data/base/repository/AbstractEntityViewAwareRepository.java
@@ -80,6 +80,18 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
     private static final String DELETE_ALL_QUERY_BY_ID_STRING = "delete from %s x where %s in :ids";
     private static final String[] EMPTY = new String[0];
     private static final EscapeCharacter DEFAULT = EscapeCharacter.of('\\');
+    private static final Pageable UNPAGED;
+
+    static {
+        Pageable unpaged = null;
+        try {
+            Method unpagedMethod = Class.forName("org.springframework.data.domain.Pageable").getMethod("unpaged");
+            unpaged = (Pageable) unpagedMethod.invoke(null);
+        } catch (Exception e) {
+            // ignore
+        }
+        UNPAGED = unpaged;
+    }
 
     protected EscapeCharacter escapeCharacter = DEFAULT;
 
@@ -327,11 +339,10 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
     }
 
     public <S extends E> Page<S> findAll(Example<S> example, Pageable pageable) {
-        ExampleSpecification<S> spec = new ExampleSpecification<>(example, escapeCharacter);
         Class<S> probeType = example.getProbeType();
         TypedQuery<S> query = getQuery(new ExampleSpecification<>(example, escapeCharacter), probeType, pageable);
 
-        return pageable == null ? new KeysetAwarePageImpl<>(query.getResultList()) : new KeysetAwarePageImpl<>((PagedList<S>) query.getResultList(), pageable);
+        return pageable == null || pageable == UNPAGED ? new KeysetAwarePageImpl<>(query.getResultList()) : new KeysetAwarePageImpl<>((PagedList<S>) query.getResultList(), pageable);
     }
 
     public List<E> findAll(Sort sort) {
@@ -339,7 +350,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
     }
 
     public Page<E> findAll(Pageable pageable) {
-        if (null == pageable) {
+        if (pageable == null || pageable == UNPAGED) {
             return (Page<E>) new PageImpl<>(findAll());
         }
 
@@ -502,7 +513,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
 
     public Page<V> findAll(Specification<E> spec, Pageable pageable) {
         TypedQuery<V> query = getQuery(spec, pageable);
-        if (pageable == null) {
+        if (pageable == null || pageable == UNPAGED) {
             return new KeysetAwarePageImpl<>(query.getResultList());
         }
         PagedList<V> resultList = (PagedList<V>) query.getResultList();
@@ -564,7 +575,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
 
         TypedQuery<V> query;
         if (entityViewClass == null) {
-            if (pageable == null) {
+            if (pageable == null || pageable == UNPAGED) {
                 query = (TypedQuery<V>) cb.getQuery();
             } else {
                 PaginatedCriteriaBuilder<S> paginatedCriteriaBuilder;
@@ -585,7 +596,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
                 query = (TypedQuery<V>) paginatedCriteriaBuilder.getQuery();
             }
         } else {
-            if (pageable == null) {
+            if (pageable == null || pageable == UNPAGED) {
                 EntityViewSetting<V, CriteriaBuilder<V>> setting = EntityViewSetting.create(entityViewClass);
                 CriteriaBuilder<V> fqb = evm.applySetting(setting, cb);
                 if (sort != null) {